refactor(Winner): Cars의 최종 우승자 관련 기능을 Winner 객체로 분리
